Update the code (but not working)
Trying to get the text to animate and then clear the background.
The text doesn't animate smoothly.
The text won't show up if resetBackground is called, anytime background(255) is called the text won't show up even if I call a delay after the text function is called and before the background function is called. WTF?