Organized and updated
Added a border / frame to organize the layout more and have a specific drawing space.
Added safety / user features to help choose colors by labeling "tap tap" - for some reason it works when it's double tapped.
Changed text on clear and submit buttons to "slide to submit" and "slide to reset" - for some reason it responds better with that. Added a 0.5 second delay after these buttons are pressed so they can't be slammed too many times.

Need to - figure out how to ask them to clear the screen after something is submitted. Arrow to the clear button, or text saying "clear screen to start over"?