Automated the file count!
Automated counting the number of files in the permanent submissions folder to create the array of projected images. Created a String array of the file names while counting the files in the folder, and then loaded the images based upon that created String & the length of the array.
Yay! So happy I figured out how to automate counting the files so that doesn't have to be updated by someone.