Life is beautiful
Everything is beautiful in the world. The sun is out and the air is crisp. The code works as it should and the birds are chirping.
Updated to make the program switch between languages.
Updated the program to be able to click properly. Something with the touchscreen monitor doesn't register mousePressed in Processing like it is supposed to. To work around, I got rid of mousePressed detection in the Button class and set isPressed to true when the mouse is over the button, reseting it to the middle of the screen after it has been "pressed" to avoid multiple presses. Adjusted the main loop to not be entirely inside mousePressed, but rather just detecting separate isPressed instances of the buttons. Changed the color to the darker blue with the color swatch for the pencil colors. Added comments and cleaned the code.